pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perturbdigger"
version = "0.1.0"
description = "Mechanism-aware perturbation prediction with partially trusted prior graphs."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
  "numpy>=1.23",
  "pandas>=1.5",
  "PyYAML>=6.0",
  "scikit-learn>=1.0",
  "scipy>=1.9",
  "h5py>=3.8",
  "tqdm>=4.66",
  "torch>=2.1"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
